refactor(authorization-server): move tokenServiceHelpers to utilities (PIN-9421)
